Reader first steps
Set up the framework for the reader. The network messager didn't work, so the reader sets up the connection. Deserializing to WMS & WFS data will be done in specific classes.
XML-files can be read from the url.
Added a custom editor script, so that url's can be evaluated outside of Play Mode. This speeds up the process a bit.