Package handling
Moved the WebReader from Assets to the existing WMSLayer package. Added functionality to request the url in-Editor, so it can be used to create examples.
Also tweaked the code to allow layers without styles to be requested (sorting them properly if there are also layers with styles. Needs to be tested still)