Feature/wms reader merged (#126)
* Reader first steps

Set up the framework for the reader. The network messager didn't work, so the reader sets up the connection. Deserializing to WMS & WFS data will be done in specific classes.
XML-files can be read from the url.
Added a custom editor script, so that url's can be evaluated outside of Play Mode. This speeds up the process a bit.

* Workflow change

Switching from using an XmlReader to read to getting the document as a string, which is then converted into an XmlDocument, allowing for more capabilities to read data from it

* WMS Formatter

Did some tweaks on the WMS Formatter and also made sure the UI works.
A setup of the WMS Formatter has been made, based on the WMS version.
Also moved some scripts to keep everything nice and uncluttered

* Web Interface

Updates to make the Web Reader's communication with the interface better

* WMS Setting

Addition of Settings to the interface.
Bounding Box can be edited, as well as the image resolution.
The reader can now correctly retrieve a map from different WMS'

* Removal of layers

Added functionality to remove layers to the reader

* Refactor and Rebuild

Refactored the code to improve data flow and redesigned the UI to fit into a side menu, just like most of the options in 3DA and 3DU

* Updated web reader

* Package handling

Moved the WebReader from Assets to the existing WMSLayer package. Added functionality to request the url in-Editor, so it can be used to create examples.
Also tweaked the code to allow layers without styles to be requested (sorting them properly if there are also layers with styles. Needs to be tested still)

* Fixed metadata issue, edited the sample scene

* Build prep

Push for building the reader manually (as autodeploy isn't working)

* Quick fix, scene error

Forgot to update the scene, thought I had already done it

* Included dependencies for building

* Code restructure (1)

* Automap preview

Code refactoring and addition of map auto-preview in the Web Service Reader

* WFS Features added

* Reordering of Web Service elements

Moved some events to reduce clutter. Moved functions from some classes to others (mostly towards the handlers)
Cleaned up the sample scene

* WFS Handling

Added MultiPoint and Polygon Handlers to handle more GeoJSON options.
Also improved WFS Interface and Handler scripts a bit

* Update ConvertCoordinates.cs

* Added LineVisualiser to NL3D

Exactly what the summary states, to be cherrypicked from the branch

* Added Handlers for WFS

Can now read WFS layers that support GeoJSON and have one of the following Geometry Types:
Point, MultiPoint, LineString, Polygon and MultiPolygon.
MultiLineString and GeometryCollection support will be added at a later time

* Setup for wfs create handler and setting parents

Can now set the parent of newly created wfs geometry or lines through an event.
Also a setup for the handler which will be responsible for creating tilehandler-based wfs layers

* Web Service Tweaks

WMS now (fairly) properly supports SRS', not just CRS'
Changed the visualiser to a default sphere to avoid issues (for the user, or with Autodesk)
WebReader scene now supports both Utrecht and Amsterdam locally

* Samples correction

* Small improvements

Tiny UI improvement, removed samples folder metadata (unwillingly pushed to the repo)

* Test Improvements

* Reworked WSR Interface

It now follows the Amsterdam and Utrecht styles of interface better. Also integrated TextMeshPro for better control over the text and it's quality.

* Update WebReader.unity

I don't know, it changed, apparently

* Added the borders for views

Added borders to the package and edited the old dual text button. Some scene changes, of which one is the re-add of the Utrecht TileHandler

* Formatter and interface update

WMS has  a new formatter, capable of filtering more layers (top level and one extra level deeper).
Amsterdam WMS & WFS data now also supported (due to less shady way of reading layers)
Separated the WMS and WFS reading features into their own interfaces and make assumptions about the link based on this.

* Coloration of WFS elements setup

* Renamed WMSLayer folder to WebServiceReader and edited the name in the package

* removed conflicting file

* removed samples folder from assets

* moved assets out of samples

* solved problem with editor trying to invoke empty events OnValidate

* changed events to be same as main

* wfs

* Corrected Invokes in the Web Service Handlers

---------

Co-authored-by: Justin <justinstolk97@gmail.com>