Conserve memory by downloading textures "non-readable"
The GetTexture method of the UnityWebRequestTexture class has an optional
flag that ensures a texture's bytes cannot be read, conserving memory
because the image is not unpacked.

In the WMSImageLayer is an option to compress an image; this does need the
texture to be readable and a special condition has been introduced to
optimize the download when it is possible.