Cleaned up 'getBAGID' and reordered fields
Part of cleaning up the getBAGID is by introducing a serialized field with the endpoint url.

Not that I expect this to change soon, but it is good practice to make these URLs configurable. This allows for consumers to introduce proxies if needed.

By adding the field above, I have also reordered the field list to group settings and events together and as such remove the need for the extra 'space' annotation.