Remove event juggling when performing a lookup
In the GeoDataLookup method, all events are removed and the default event is reattached. This is a potential bug as any events attached by a caller are removed as well. This juggling is done to prevent the suggestions from triggering when setting the text, but there is a nifty 'SetTextWithoutNotify' method for this use case.

By using SetTextWithoutNotify, we can remove the event juggling and make this a bit more solid.