Move Core/Tilesystem into dedicated package
As an example how we can split the Netherlands3D package into smaller packages; I have extracted the Core/Tilesystem package from the Netherlands3D package. With this change we can actually publish smaller packages, aiding in documenting and publishing these components separately.

An interesting issue that I did run into is that you cannot make dependencies from a package to a git location; this would mean that you a user is always responsible for importing all packages themselves, including dependencies.

This can be solved by publishing the packages to a registry or by pushing them to the asset store; but we need to discuss how to continue with this.