Introduce the Authentication package
In this package, I have introduced the concept of a Session. A Session represents the means to sign in into a service, such as Azure AD. With a session you do not need to be signed -i.e. anonymous-.

The Session is a ScriptableObject so that it can be re-used across scenes; and a SessionListener component is introduced to listen to the transitions that can happen.

This package relies on the CDMVision's Authentication package; which provides a good OAuth client that we wrap to make it easier to use; and to introduce an Azure AD version of the OAuth client.