Revert pipeline changes
Apparently these pipeline changes give weird side-effects, so I have reverted it to prevent side effects in my current PR. If this still proves to be troublesome in newer Unity versions we should address this separately.

This has mostly to do with the loading of WMS overlays, they are not shown or the building cut-out option does not work. This looks to be related to changes in Unity how it works with the render pipelines