Make ExampleMunicipality great again!
Aside from the corny intro to this change, what I have done is replace all references to local paths to the new online environment for Flevoland. With this change, the example is immediately usable and it is no longer needed to make sure you have tiles from Flevoland built on your local machine.

In addition, the Graphics Settings did not refer to the Render Pipeline; and by adding that it will also help make the example immediately usable.

It might be worth noting that the pathQuery provided in the Buildings and Terrain gameobjects is only valid for one year; so if it stops working: you should probably replace the token.