Move Event System into its own Assembly Definition
In an effort to separate the Core package into independent pieces and, if all is well, have a TileHandler package left; I have split off the event system from the Core package.

This will also visualize which packages depend on the event system, and thus which we may want to refactor in order to separate the event system from the packages that we have.

The affected Assembly Definitions, and thus packages that use the event system, are:

- Address Search
- BAGInformation
- BIMPlanning
- Cameras
- Core
- File Importer
- GetWMSUrl
- ModelParsing
- ProfileRendering
- SelectionTools
- SubObjectFiltering
- Sun
- Timeline
- Traffic
- Visualisers
- WebServiceReader
- Windmills

All other packages are free of the event system and could possibly be migrated to real packages more easily.