Extract Snapshot to its own package
In this change, I have extracted the snapshot package to its own location. I did go a bit overboard and updated the package and fixed a couple of bugs.

Especially the introduction of the Snapshot static class with the various helper methods is going to help greatly in creating images to use in reports; such as the shadow report.