Remove the old Obj Importer code and replace it with the new package
In this change, a new dependency has been set to the new
eu.netherlands3d.obj-importer package; and all code that has been extracted
to this package is removed. All code paths depending on the old code have
been checked and updated where it was needed due to the introduction of
namespaces in that package.