Introduce Web pseudo-package
In this change, I have moved all assisting functionality for WebRequests into
a single pseudo-package "Web". This functionality will make the
WebServiceReader pseudo-package lighter and adds the parsing of Query
Parameters from Query Strings.

For now, it has not yet been elevated to a true package because we can have
discuss the next steps based on this.